alpr/requirements.txt
Cython==0.29.23
numpy==1.20.3
opencv-python==4.5.2.52
python-dotenv==0.17.1
